Backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.0
pydantic>=2.6.0
pydantic-settings>=2.0.0
alembic>=1.13.0
azure-identity>=1.16.0
azure-mgmt-appcontainers>=3.0.0
psycopg[binary]>=3.1.0
python-dotenv>=1.0.0
